pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-latest-development"
version = "0.1.0"
description = "Gmail automation and AI agent utilities"
authors = [
  { name = "Dhruv Patel", email = "dhruvpatel@example.com" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "python-dotenv",
  "google-api-python-client",
  "google-auth",
  "google-auth-oauthlib",
  "google-auth-httplib2",
  "requests"
]

[tool.setuptools.packages.find]
where = ["src"]
